moved 'OpenPypeInterface' into interfaces.py
